refactor: refactor structure a little
