+++ index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.ico" />
+    <link rel="icon" type="image/svg+xml" href="/images/vite.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>Graphiql-app</title>
   </head>
