+++ src/components/schemaViewer/TypeTab/TabHeader.tsx
@@ -4,11 +4,7 @@ import { isOutputFieldType, isScalarFieldType } from '../utils';
 import { FieldInfo } from '../ui/FieldInfo';
 import { TypeInfo } from '../ui/TypeInfo';
 
-type TabHeaderProps = {
-  typeToDisplay: TypeToDisplay;
-};
-
-export function TabHeader({ typeToDisplay }: TabHeaderProps) {
+export function TabHeader({ typeToDisplay }: { typeToDisplay: TypeToDisplay }) {
   const tabHeaderWithoutArgs = (
     <TypeInfo name={typeToDisplay.name} type={typeToDisplay.type.toString()} />
   );
